Fix toolbox not initializing due to enabledChanged signal
The Connections in qml is a QObject and it has an enabledChanged signal by default.
Therefore, we cannot really use the Connections to connect to the enabledChanged signal of e.g.
a python class, because qml complains that there is already a function named enabledChanged() in
the Connections.

To circumvent that, we can simply rename the enabledChanged() signal of the Toolbox to
toolboxEnabledChanged().